Create codeql.yml
Invsioning future learning curve, that have many supplements to a lose of dynamics on this project. I had to be heard or it was already in the works... I'm just a messenger.  Trying to push the limitless Artifact Intellectuals buttons for the greater good to see light even when humanity is at the darkest of time. Direct to main branch ... Please